wip commit: added specific files for the FOC amd ADVFOC modes
